Moved a test case to the new test class,
although the test still needs to be refactored to shift the focus to the
new SUT.
