+++ AtomEventStore.UnitTests/XmlAtomContentTests.cs
@@ -145,31 +145,6 @@ namespace Grean.AtomEventStore.UnitTests
         }
 
         [Theory, AutoAtomData]
-        public void SutCanSerializeNestedItem(
-            XmlAtomContent seed,
-            Envelope<TestEventX> env)
-        {
-            var sut = seed.WithItem(env);
-
-            var actual = sut.ToXmlString(
-                new ConventionBasedSerializerOfComplexImmutableClasses());
-
-            var expected = XDocument.Parse(
-                "<content type=\"application/xml\" xmlns=\"http://www.w3.org/2005/Atom\">" +
-                "  <envelope xmlns=\"urn:grean:atom-event-store:unit-tests\">" +
-                "    <id>urn:uuid:" + env.Id + "</id>" +
-                "    <item>" +
-                "      <test-event-x>" +
-                "        <number>" + env.Item.Number + "</number>" +
-                "        <text>" + env.Item.Text + "</text>" +
-                "      </test-event-x>" +
-                "    </item>" +
-                "  </envelope>" +
-                "</content>");
-            Assert.Equal(expected, XDocument.Parse(actual), new XNodeEqualityComparer());
-        }
-
-        [Theory, AutoAtomData]
         public void SutCanRoundTripNestedItem(
             XmlAtomContent seed,
             Envelope<TestEventY> env)
